requirements.txt
numpy
sklearn
tensorflow
pillow
requests
